Rename ModelInoutSamples -> WidhetInputSamples
