+++ widgets/src/lib/InferenceWidget/shared/WidgetInputSamples/WidgetInputSamples.svelte
